Extract usePipelineEditing hook from PipelineTable and KanbanBoard
